requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4
PyJWT==2.8.0
python-dotenv==1.0.1
